+++ nvim/lua/custom/plugins/pr-review.lua
@@ -410,11 +410,17 @@ return {
           review_diff = {
             select_next_entry = { lhs = ']f', desc = 'move to next changed file' },
             select_prev_entry = { lhs = '[f', desc = 'move to previous changed file' },
+            toggle_viewed = { lhs = '<Tab>', desc = 'toggle file viewed' },
+            select_next_unviewed_entry = { lhs = ']u', desc = 'next unviewed file' },
+            select_prev_unviewed_entry = { lhs = '[u', desc = 'prev unviewed file' },
           },
           file_panel = {
             select_next_entry = { lhs = ']f', desc = 'move to next changed file' },
             select_prev_entry = { lhs = '[f', desc = 'move to previous changed file' },
             select_entry = { lhs = '<CR>', desc = 'open file' },
+            toggle_viewed = { lhs = '<Tab>', desc = 'toggle file viewed' },
+            select_next_unviewed_entry = { lhs = ']u', desc = 'next unviewed file' },
+            select_prev_unviewed_entry = { lhs = '[u', desc = 'prev unviewed file' },
           },
           review_thread = {
             select_next_entry = { lhs = ']f', desc = 'move to next changed file' },